+++ modules/optics/arrow-optics/src/main/kotlin/arrow/optics/Getter.kt
@@ -8,7 +8,6 @@ import arrow.core.Tuple2
 import arrow.core.compose
 import arrow.core.identity
 import arrow.core.toT
-import arrow.mtl.map
 import arrow.higherkind
 import arrow.typeclasses.Monoid
